+++ packages/spacecat-shared-google-client/CHANGELOG.md
@@ -1,3 +1,9 @@
+## [@adobe/spacecat-shared-google-client-v1.5.8](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-google-client-v1.5.7...@adobe/spacecat-shared-google-client-v1.5.8) (2026-03-28)
+
+### Bug Fixes
+
+* **deps:** update external fixes ([#1477](https://github.com/adobe/spacecat-shared/issues/1477)) ([67bdd1a](https://github.com/adobe/spacecat-shared/commit/67bdd1a2c497bed088bc1e54ae22e60c171308d1))
+
 ## [@adobe/spacecat-shared-google-client-v1.5.7](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-google-client-v1.5.6...@adobe/spacecat-shared-google-client-v1.5.7) (2026-03-21)
 
 ### Bug Fixes